Added several more explicit casts.
Affected modules: misc scheduling/flow sim

+++ src/misc/string_utils.cc
@@ -19,13 +19,13 @@ void spf(string* s, const std::string fmt, ...) {
   va_list marker;
 
   while (!b) {
-    s->resize(size);
+    s->resize(static_cast<size_t>(size));
     va_start(marker, fmt);
     // N.B. the below relies on a slightly dodgy cast via c_str()!
     n = vsnprintf((char*)s->c_str(), size, fmt.c_str(), marker);  // NOLINT -- deconstify
     va_end(marker);
     if ((n > 0) && ((b = (n < size)) == true))
-      s->resize(n);
+      s->resize(static_cast<size_t>(n));
     else
       size *= 2;
   }
@@ -39,13 +39,13 @@ void spfa(string *s, const std::string fmt, ...) {
   va_list marker;
 
   while (!b) {
-    ss.resize(size);
+    ss.resize(static_cast<size_t>(size));
     va_start(marker, fmt);
     // N.B. the below relies on a slightly dodgy cast via c_str()!
     n = vsnprintf((char*)ss.c_str(), size, fmt.c_str(), marker);  // NOLINT -- deconstify
     va_end(marker);
     if ((n > 0) && ((b = (n < size)) == true))
-      ss.resize(n);
+      ss.resize(static_cast<size_t>(n));
     else
       size *= 2;
   }
